+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 - [__change__:] Plugin requires Android KitKat or newer
 - [__change__:] `isAvailable` returns false if no enabled service can be found (Android)
 - [feature:] New `pick` interface to pick a printer for future usage
+- [feature:] Initial support for Windows10 platform (Thanks to #cristi-badila)
 - [enhancement:] `isAvailable` returns count of available services (Android)
 - [enhancement:] `print` returns bool value to indicate the result
 - [enhancement:] Added missing `duplex` support (Android)
 - [__change__:] `duplex` requires a string (`none`, `long` or `short`)
-- [enhancement:] Support for `hidePageRange`, `hideNumberOfCopies` and `hidePaperFormat` (iOS specific)
+- [enhancement:] Support for `border`, `hidePageRange`, `hideNumberOfCopies` and `hidePaperFormat` (iOS specific)
 
 
 #### Version 0.7.1 (23.04.2015)